fixed up some error types, and signature errors
