+++ test/components/util/state-machine.test.tsx
@@ -25,6 +25,8 @@ import {fromMessage} from "../../../src/util/messages";
 import Wave from "../../../src/models/wave";
 import MessageSender = chrome.runtime.MessageSender;
 import Options from "../../../src/models/options";
+import {log, ComponentLog, MachineComponentProps, ReactMachine, UseStateProxy} from "../../../src/util/react-machine";
+
 
 const BaseVentures = ["base", "error"]
 
@@ -182,199 +184,7 @@ describe("state machine", () => {
         expect((await stateMachine.handleState(stateNameMap.getState("selection mode deactivate")!!))?.name).toBe("base");
     })
 
-    type MachineComponentProps = {
-        stateProxy: UseStateProxy
-        machine: StateMachine
-        previousState: string
-    }
-    type MachineComponent = ReactElement<any, any> | null & {
-    }
 
-    type ComponentLog = {
-        state: string,
-        view?: MachineComponent
-    }
-    type ComponentLogView = {
-        states: string[]
-        state: string
-        views: MachineComponent[]
-    }
-    const log = (state: string, view?: MachineComponent) => {
-        return {
-            state,
-            view
-        } as unknown as ComponentLog
-    }
-    const view = (state: string, states: string[], ...views: MachineComponent[]) => {
-        return {
-            state,
-            states,
-            views
-        } as unknown as ComponentLogView
-    }
-
-    type StateComponentFunction = { ({stateProxy, machine}: Partial<MachineComponentProps>): Promise<ComponentLog> }
-    type ReactStateMachineProps = {
-        states: { [key: string]: StateComponentFunction },
-        client: Client<Message<any>>,
-        useStateProxy: UseStateProxy,
-        errorState?: State
-    }
-
-    type useStateFunction<S> = { (initialState: S | (() => S)): [S, Dispatch<SetStateAction<S>>] }
-
-    class UseStateProxy {
-        useStateFn?: useStateFunction<any>
-        useStates: Map<string, SetStateAction<any>> = new Map<string, React.SetStateAction<any>>()
-        stateValues: Map<string, any> = new Map<string, any>()
-
-        constructor(useStateFn?: useStateFunction<any>) {
-            this.useStateFn = useStateFn
-        }
-
-        setUseState(useStateFn: useStateFunction<any>) {
-            this.useStateFn = useStateFn
-        }
-        useState<S>(name: string, initialState: S | (() => S)): [S, Dispatch<SetStateAction<S>>] {
-            const [state, setState] = this.useStateFn!(initialState)
-            return [state, ((value: S): void => {
-                this.stateValues.set(name, value)
-                setState(value);
-            }) as Dispatch<SetStateAction<S>>]
-        }
-        getState<T>(name: string): T {
-            return this.stateValues.get(name) as T;
-        }
-    }
-
-    type whatever = null | undefined
-
-    type ReactStateMachineComponentProps = {
-        useState: useStateFunction<any>
-        setState: { (state: string, value: any): void }
-    }
-
-    class ReactStateMachine {
-        private states: { [key: string]: StateComponentFunction };
-        private renderTarget?: React.ReactElement<any, any> | null;
-        private stateMachine: StateMachine = new StateMachine()
-        private logView: ComponentLogView = {
-            state: "base",
-            views: [],
-            states: ["base"]
-        }
-        private stateProxy: UseStateProxy = new UseStateProxy()
-        private errorView: string[] = [];
-        private ErrorStateFunction = (): ComponentLog => {
-            return log("base", <ul>{this.errorView.map(e => <li>{e}</li>)}</ul>);
-        }
-
-        private ErrorState = new State("error", ["error", "base"], true, (async (message, state, previousState): State => {
-            const log = this.ErrorStateFunction()
-            if (log.view) this.logView.views.push(log.view)
-            this.logView.states.push(log.state)
-            return this.stateMachine.getState((await log).state) || this.ErrorState
-        }))
-
-        private errorState = this.ErrorState;
-
-        private toState(name: string, isBaseLevel: boolean, ventureStates: string[], componentFunction: StateComponentFunction) {
-            return new State(name, ventureStates, isBaseLevel, (async (message, state, previousState): State => {
-                const log: ComponentLog = await componentFunction({
-
-                } as Partial<MachineComponentProps>)
-                if (log.view) this.logView.views.push(log.view)
-                this.logView.states.push(log.state)
-                return this.stateMachine.getState((await log).state) || this.ErrorState
-            }))
-        }
-
-        constructor({
-            states,
-            errorState
-        }: ReactStateMachineProps) {
-            this.states = states;
-            this.errorState = errorState || this.ErrorState;
-        }
-
-        initialize() {
-            const machine = this;
-            const map = new Map<string, State>();
-
-            if (!(this.logView.state in this.states)) {
-                machine.errorView.push("cannot find " + machine.logView.state + " in states map!")
-            }
-
-            this.stateMachine.initialize(new class implements NameAccessMapInterface {
-                getState(name: string): State | undefined {
-                    return map.get(name);
-                }
-            }, map.get(this.logView.state) || this.ErrorState)
-
-            this.renderTarget = ReactStateMachineRenderTarget({
-                useStateProxy: this.stateProxy,
-                logView: this.logView,
-                stateMachine: this.stateMachine
-            })
-        }
-
-        handleMessage(message: Named) {
-            this.stateMachine.handleState(message);
-        }
-
-        getStateMachine(): StateMachine {
-            return this.stateMachine;
-        }
-
-        getLogView(): ComponentLogView {
-            return this.logView;
-        }
-
-        getRenderTarget(): ReactElement<any, any> | whatever {
-            return this.renderTarget;
-        }
-    }
-
-
-    type ReactStateMachineRenderTargetProps = {
-        useStateProxy: UseStateProxy,
-        logView: ComponentLogView,
-        stateMachine: StateMachine,
-        context?: React.Context<StateMachine>
-    }
-
-    const ReactStateMachineRenderTarget: FunctionComponent<ReactStateMachineRenderTargetProps> = ({
-        useStateProxy,
-        logView,
-        stateMachine,
-        context= React.createContext(stateMachine)
-    }): ReactElement<any, any> | null => {
-        const [rerender, setRerender] = useState(5);
-
-        useEffect(() => {
-            useStateProxy.setUseState(useState)
-
-            // todo: review: we may want to subscribe from the componentMachine
-            stateMachine.getObservable()?.subscribe(() => {
-                setRerender(rerender + 1);
-            })
-        }, [])
-
-
-        useEffect(() => {
-            if (rerender >= 100) {
-                setRerender(1);
-            }
-        }, [rerender])
-
-        return (<context.Provider value={stateMachine}>
-            {logView.views.map((view, i) => <view key={i} />)}
-        </context.Provider>);
-    }
-
-    const ReactMachine = (props: ReactStateMachineProps) => {
-        return new ReactStateMachine(props)
-    }
 
     class TestRuntimeProxy implements IRuntimeProxy {
         public reponses: any[] = []
@@ -418,6 +228,7 @@ describe("state machine", () => {
             states: {
                 base: ({stateProxy, machine}: Partial<MachineComponentProps>): Promise<ComponentLog> => {
                     const [test, setTest] = stateProxy?.useState("test", false) || [undefined, undefined]
+                    // todo: review: replace log and view with a deconstructor like useState and useReducer?
                     return Promise.resolve(log("base", <div/>));
                 },
                 complex: async ({stateProxy, machine}: Partial<MachineComponentProps>): Promise<ComponentLog> => {
